Rename helpers file to proper naming
