+++ test/lib/helpers.js
